Доработка

+++ project/apps/posts/src/app/editing/editing.service.ts
@@ -1,9 +1,9 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { BlogPostRepository } from '../blog-post/blog-post.repository';
 import { BlogTagRepository } from '../blog-tag/blog-tag.repository';
-import { CreatePostDto } from './dto/create-post.dto';
-import { POST_NOT_FOUND, Post } from '@project/shared/shared-types';
+import { POST_NOT_FOUND, Post, PostState } from '@project/shared/shared-types';
 import { BlogPostEntity } from '../blog-post/blog-post.entity';
+import { CreatePostDto, EditPostDto } from '@project/shared/shared-dto';
 
 
 @Injectable()
@@ -15,11 +15,16 @@ export class EditingService {
 
   async createPost(dto: CreatePostDto): Promise<Post> {
     const tags = await this.blogTagRepository.find(dto.tags);
-    const postEntity = new BlogPostEntity({ ...dto, tags, comments: [], commentsCount: 0, likesCount: 0 });
+    const postEntity = new BlogPostEntity({ ...dto, tags, comments: [], commentsCount: 0, likesCount: 0, postState: PostState.Public });
     return this.blogPostRepository.create(postEntity);
   }
 
   async deletePostId(id: number): Promise<void> {
+    const existPost = await this.blogPostRepository.findById(id);
+
+    if (!existPost) {
+      throw new NotFoundException(POST_NOT_FOUND);
+    }
     this.blogPostRepository.destroy(id);
   }
 
@@ -32,18 +37,7 @@ export class EditingService {
     return this.blogPostRepository.findById(id);
   }
 
-  async getPostTitle(title: string) {
-
-    const existPost = await this.blogPostRepository.findByTitle(title);
-
-    if (!existPost) {
-      throw new NotFoundException(POST_NOT_FOUND);
-    }
-    return this.blogPostRepository.findByTitle(title);
-  }
-
-
-  async updatePostId(id: number, dto: CreatePostDto) {
+  async updatePostId(id: number, dto: EditPostDto) {
     const tags = await this.blogTagRepository.find(dto.tags);
     const existPost = await this.blogPostRepository.findById(id);
 
@@ -73,6 +67,7 @@ export class EditingService {
           ...existPost,
           isRepost: true,
           originUserId: existPost.userId,
+          originPostId: id,
           userId: newUserId};
 
     const postEntity = new BlogPostEntity(blogPost);
